// packages/moim-web/app/actions/types.tsx
export enum GroupTypes {
  SUCCEEDED_GET_INITIAL_GROUP_ID = "GROUP.SUCCEEDED_GET_INITIAL_GROUP_ID",

  START_BOOT_STRAP = "GROUP.START_BOOT_STRAP",
  SUCCESS_BOOT_STRAP = "GROUP.SUCCESS_BOOT_STRAP",
  FAILED_BOOT_STRAP = "GROUP.FAILED_BOOT_STRAP",

  START_VALIDATE_GROUP_DOMAIN = "GROUP.START_VALIDATE_GROUP_DOMAIN",
  VALIDATE_GROUP_DOMAIN = "GROUP.VALIDATE_GROUP_DOMAIN",

  START_FETCHING_LIST_GROUP = "GROUP.START_FETCHING_LIST_GROUP",
  FETCHED_LIST_GROUP = "GROUP.FETCHED_LIST_GROUP",
  FAILED_FETCHING_LIST_GROUP = "GROUP.FAILED_FETCHING_LIST_GROUP",

  START_UPDATE_GROUP_ICON = "GROUP.START_UPDATE_GROUP_ICON",
  SUCCEED_UPDATE_GROUP_ICON = "GROUP.SUCCEED_UPDATE_GROUP_ICON",
  FAILED_UPDATE_GROUP_ICON = "GROUP.FAILED_UPDATE_GROUP_ICON",

  START_CROP_GROUP_ICON = "GROUP.START_CROP_GROUP_ICON",
  SUCCEEDED_CROP_GROUP_ICON = "GROUP.SUCCEEDED_CROP_GROUP_ICON",
  FAILED_CROP_GROUP_ICON = "GROUP.FAILED_CROP_GROUP_ICON",

  START_CROP_GROUP_BANNER = "GROUP.START_CROP_GROUP_BANNER",
  SUCCEEDED_CROP_GROUP_BANNER = "GROUP.SUCCEEDED_CROP_GROUP_BANNER",
  FAILED_CROP_GROUP_BANNER = "GROUP.FAILED_CROP_GROUP_BANNER",

  START_UPDATE_GROUP_BANNER = "GROUP.START_UPDATE_GROUP_BANNER",
  SUCCEEDED_UPDATE_GROUP_BANNER = "GROUP.SUCCEEDED_UPDATE_GROUP_BANNER",
  FAILED_UPDATE_GROUP_BANNER = "GROUP.FAILED_UPDATE_GROUP_BANNER",

  START_CREATE_GROUP = "GROUP.START_CREATE_GROUP",
  SUCCEEDED_CREATE_GROUP = "GROUP.SUCCEEDED_CREATE_GROUP",
  FAILED_CREATE_GROUP = "GROUP.FAILED_CREATE_GROUP",

  START_CREATE_SUB_GROUP = "GROUP.START_CREATE_SUB_GROUP",
  SUCCEEDED_CREATE_SUB_GROUP = "GROUP.SUCCEEDED_CREATE_SUB_GROUP",
  FAILED_CREATE_SUB_GROUP = "GROUP.FAILED_CREATE_SUB_GROUP",

  START_GET_GROUP_DATA = "GROUP.START_GET_GROUP_DATA",
  SUCCEEDED_GET_GROUP_DATA = "GROUP.SUCCEEDED_GET_GROUP_DATA",
  FAILED_GET_GROUP_DATA = "GROUP.FAILED_GET_GROUP_DATA",

  START_FETCHING_GROUP_STATUS = "GROUP.START_FETCHING_GROUP_STATUS",
  FETCHED_GROUP_STATUS = "GROUP.FETCHED_GROUP_STATUS",
  FAILED_FETCHING_GROUP_STATUS = "GROUP.FAILED_FETCHING_GROUP_STATUS",

  START_FETCHING_SUB_MOIMS = "GROUP.START_FETCHING_SUB_MOIMS",
  FETCHED_SUB_MOIMS = "GROUP.FETCHED_SUB_MOIMS",
  FAILED_FETCHING_SUB_MOIMS = "GROUP.FAILED_FETCHING_SUB_MOIMS",

  START_GET_JOINED_SUB_MOIMS = "GROUP.START_GET_JOINED_SUB_MOIMS",
  SUCCEED_GET_JOINED_SUB_MOIMS = "GROUP.SUCCEED_FETCHED_JOINED_SUB_MOIMS",
  FAILED_GET_JOINED_SUB_MOIMS = "GROUP.FAILED_GET_JOINED_SUB_MOIMS",

  OPEN_JOIN_GROUP_DIALOG = "GROUP.OPEN_JOIN_GROUP_DIALOG",
  CLOSE_JOIN_GROUP_DIALOG = "GROUP.CLOSE_JOIN_GROUP_DIALOG",

  CLEAR_MOIM_ICON = "GROUP.CLEAR_MOIM_ICON",

  START_RENAME_MOIM = "GROUP.START_RENAME_MOIM",
  SUCCEED_RENAME_MOIM = "GROUP.SUCCEED_RENAME_MOIM",
  FAILED_RENAME_MOIM = "GROUP.FAILED_RENAME_MOIM",

  START_SET_DESCRIPTION_MOIM = "GROUP.START_SET_DESCRIPTION_MOIM",
  SUCCEED_SET_DESCRIPTION_MOIM = "GROUP.SUCCEED_SET_DESCRIPTION_MOIM",
  FAILED_SET_DESCRIPTION_MOIM = "GROUP.FAILED_SET_DESCRIPTION_MOIM",

  START_GET_MOIM_THEME = "GROUP.START_GET_MOIM_THEME",
  SUCCEED_GET_MOIM_THEME = "GROUP.SUCCEED_GET_MOIM_THEME",
  FAILED_GET_MOIM_THEME = "GROUP.FAILED_GET_MOIM_THEME",

  START_GET_MOIM_COVER = "GROUP.START_GET_MOIM_COVER",
  SUCCEED_GET_MOIM_COVER = "GROUP.SUCCEED_GET_MOIM_COVER",
  FAILED_GET_MOIM_COVER = "GROUP.FAILED_GET_MOIM_COVER",

  START_UPDATE_HOME_CHANNEL = "GROUP.START_UPDATE_DEFAULT_CHANNEL",
  SUCCEED_UPDATE_HOME_CHANNEL = "GROUP.SUCCEED_UPDATE_DEFAULT_CHANNEL",
  FAILED_UPDATE_HOME_CHANNEL = "GROUP.FAILED_UPDATE_DEFAULT_CHANNEL",

  START_GET_PARENT_THEME = "GROUP.START_GET_PARENT_THEME",
  SUCCEED_GET_PARENT_THEME = "GROUP.SUCCEED_GET_PARENT_THEME",
  FAILED_GET_PARENT_THEME = "GROUP.FAILED_GET_PARENT_THEME",

  START_GET_RECOMMEND_MOIMS = "GROUP.START_GET_RECOMMEND_MOIMS",
  SUCCEED_GET_RECOMMEND_MOIMS = "GROUP.SUCCEED_GET_RECOMMEND_MOIMS",
  FAILED_GET_RECOMMEND_MOIMS = "GROUP.FAILED_GET_RECOMMEND_MOIMS",

  START_SEARCH_MOIMS = "GROUP.START_SEARCH_MOIMS",
  SUCCEED_SEARCH_MOIMS = "GROUP.SUCCEED_SEARCH_MOIMS",
  FAILED_SEARCH_MOIMS = "GROUP.FAILED_SEARCH_MOIMS",
  CLEAR_SEARCH_MOIMS = "GROUP.CLEAR_SEARCH_MOIMS",

  START_GET_INSTALLED_PLUGINS = "GROUP.START_GET_INSTALLED_PLUGINS",
  SUCCEED_GET_INSTALLED_PLUGINS = "GROUP.SUCCEED_GET_INSTALLED_PLUGINS",
  FAILED_GET_INSTALLED_PLUGINS = "GROUP.FAILED_GET_INSTALLED_PLUGINS",
}
// ...
export enum AppTypes {
  CURRENT_GROUP_CHANGED = "APP.CURRENT_GROUP_CHANGED",
  PARENT_GROUP_CHANGED = "PARENT_GROUP_CHANGED",

  SIGN_OUT = "SIGN_OUT",
  CHANGE_LOCALE = "CHANGE_LOCALE",

  SUCCEED_USER_BLOCK = "APP.SUCCEED_USER_BLOCK",
  SUCCEED_USER_UNBLOCK = "APP.SUCCEED_USER_UNBLOCK",

  USER_GLOBAL_MUTE_NOTIFICATION = "APP.USER_GLOBAL_MUTE_NOTIFICATION",

  INITIALIZE_HISTORY = "INITIALIZE_HISTORY",
  PUSH_HISTORY = "PUSH_HISTORY",

  LOCATION_CHANGE = "@@router/LOCATION_CHANGE",
  START_REFRESH_TOKEN = "APP.START_REFRESH_TOKEN",
  SUCCEED_REFRESH_TOKEN = "APP.SUCCEED_REFRESH_TOKEN",
  FAILED_REFRESH_TOKEN = "APP.FAILED_REFRESH_TOKEN",

  SET_PROVIDERS = "APP.SET_PROVIDERS",
  SET_APP_CONFIG = "APP.SET_APP_CONFIG",
}

export enum AuthTypes {
  START_REFRESH_TOKEN = "APP.START_REFRESH_TOKEN",
  SUCCEED_REFRESH_TOKEN = "APP.SUCCEED_REFRESH_TOKEN",
  FAILED_REFRESH_TOKEN = "APP.FAILED_REFRESH_TOKEN",

  START_LOGGING_IN = "START_LOGGING_IN",
  END_LOGGING_IN = "END_LOGGING_IN",
  SIGN_OUT = "SIGN_OUT",

  CURRENT_USER_CHANGED = "CURRENT_USER_CHANGED",
  PARENT_MOIM_USER_CHANGED = "PARENT_MOIM_USER_CHANGED",

  START_GET_AUTHENTICATION = "APP.START_GET_AUTHENTICATION",
  SUCCEED_GET_AUTHENTICATION = "APP.SUCCEED_GET_AUTHENTICATION",
  FAILED_GET_AUTHENTICATION = "APP.FAILED_GET_AUTHENTICATION",
}